+++ src/components/home-container/index.jsx
@@ -4,7 +4,7 @@ import styles from "components/home-container/index.module.scss";
 import Button from "components/base/button";
 import CreateTask from "components/task/create-new";
 import TaskCard from "components/task/existing-task/task-card";
-import FilterBtnContainer from "components/filter-btn-container";
+import FilterButtonGroup from "components/filter-button-group";
 
 function HomeContainer() {
   const tasks = useSelector((state) => state.todo.tasks);
@@ -23,7 +23,7 @@ function HomeContainer() {
       <h1>Add Tasks</h1>
       <div className={`flex justify-between ${styles.actionBarContainer}`}>
         <Button onClick={showNewTaskCard}>Create</Button>
-        <FilterBtnContainer />
+        <FilterButtonGroup />
       </div>
 
       <div className="grid grid-cols-3 card-gap">
